+++ base/ViewEvent.php
@@ -23,24 +23,9 @@ class ViewEvent extends Event
 	 */
 	public $output;
 	/**
-	 * @var string the view file path that is being rendered by [[View::renderFile()]].
-	 */
-	public $viewFile;
-	/**
 	 * @var boolean whether to continue rendering the view file. Event handlers of
 	 * [[View::EVENT_BEFORE_RENDER]] may set this property to decide whether
 	 * to continue rendering the current view file.
 	 */
 	public $isValid = true;
-
-	/**
-	 * Constructor.
-	 * @param string $viewFile the view file path that is being rendered by [[View::renderFile()]].
-	 * @param array $config name-value pairs that will be used to initialize the object properties
-	 */
-	public function __construct($viewFile, $config = [])
-	{
-		$this->viewFile = $viewFile;
-		parent::__construct($config);
-	}
 }